✨ Add Enumeration base class for domain entities
This commit adds a new base class `Enumeration` to the `Domain.Shared` namespace. The `Enumeration` class is an abstract base class for enumerations and provides common functionality for working with enumeration values. It includes methods for retrieving enumeration values based on their ID or name, as well as retrieving all enumeration values. This base class will be used to define specific enumerations in the domain.